server/requirements.txt
# -----------------------------
# Core API
# -----------------------------
fastapi
uvicorn[standard]
python-dotenv

# -----------------------------
# Machine Learning
# -----------------------------
numpy
pandas
scikit-learn
joblib

# -----------------------------
# Utilities
# -----------------------------
pydantic
